Scripts/requirements.txt
soundfile>=0.12.0
numpy>=1.20.0
scipy>=1.7.0
Pillow>=9.0.0
